+++ fractal.h
@@ -9,10 +9,10 @@
 class Fractal {
 	FractalParams params;
 	ValueProvider* valueProvider;
-	inline void increaseValueAt(long** image, const Complex& complex);
-	void addToBmp(long** image, Bmp& bmpImage, const Vector& multiplier, double colorMultiplier);
-	void deleteImage(long** image);
-	long** drawInternal(int index, long long iterationLimit, long long iterationCount, long long minIterationCount);
+	inline void increaseValueAt(int** image, const Complex& complex);
+	void addToBmp(int** image, Bmp& bmpImage, const Vector& multiplier, double colorMultiplier);
+	void deleteImage(int** image);
+	int** drawInternal(int index, long long iterationLimit, long long iterationCount, long long minIterationCount);
 
 	public:
 		Fractal(ValueProvider* valueProvider, const FractalParams& params);
